clean up after merge
